# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MSC Backend API Documentation",
        "description": "Interactive Swagger API documentation for MSC Backend",
        "contact": {
            "name": "MSC Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/manager/employees/bulk/validate": {
            "post": {
                "tags": [
                    "Manager Management"
                ],
                "summary": "Validate bulk employee upload",
                "operationId": "f7b4ede300448dd05b4e275646d7200b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "CSV or XLSX file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation results"
                    },
                    "422": {
                        "description": "Invalid file format or validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/employees/bulk": {
            "post": {
                "tags": [
                    "Manager Management"
                ],
                "summary": "Process bulk employee upload",
                "operationId": "048aba27a46b2cd1857582e687fe6b7b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "CSV or XLSX file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Upload successful"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/employees/bulk/template": {
            "get": {
                "tags": [
                    "Manager Management"
                ],
                "summary": "Download bulk employee upload template",
                "operationId": "d88cc6d08051638cf41dfe82ddfe6a14",
                "responses": {
                    "200": {
                        "description": "Template downloaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/employees": {
            "get": {
                "tags": [
                    "Manager Management"
                ],
                "summary": "List all employees under the authenticated manager",
                "operationId": "8817bd43af00e162bb50ee092ab16af4",
                "responses": {
                    "200": {
                        "description": "List of employees"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/employees/{id}": {
            "get": {
                "tags": [
                    "Manager Management"
                ],
                "summary": "Get a specific employee's details",
                "operationId": "18ee685b5fe254a61e0aab32db07288b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized access to this employee"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Manager Management"
                ],
                "summary": "Update a specific employee's details",
                "operationId": "1b7f63871f0569d0bfb2fe9477d7c78a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "department_role": {
                                        "type": "string"
                                    },
                                    "branch": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized access to this employee"
                    },
                    "404": {
                        "description": "Employee not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Manager Management"
                ],
                "summary": "Deactivate an employee (soft delete)",
                "operationId": "2f9106f79a464b16ac0b856112b23414",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee deactivated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized access to this employee"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter ONLY your personal access token (e.g., 1|xxxx...)",
                "scheme": "bearer"
            }
        }
    },
...
}